==> Project Folder/1. Clean Data.ipynb <==
<h1>Tag Cleaning<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Loading-Required-libraries" data-toc-modified-id="Loading-Required-libraries-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Loading Required libraries</a></span></li><li><span><a href="#Tag-Cleaning" data-toc-modified-id="Tag-Cleaning-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Tag Cleaning</a></span><ul class="toc-item"><li><span><a href="#Attribute-Name" data-toc-modified-id="Attribute-Name-2.1"><span class="toc-item-num">2.1&nbsp;&nbsp;</span>Attribute Name</a></span><ul class="toc-item"><li><span><a href="#Check-and-Clean-Tag" data-toc-modified-id="Check-and-Clean-Tag-2.1.1"><span class="toc-item-num">2.1.1&nbsp;&nbsp;</span>Check and Clean Tag</a></span></li><li><span><a href="#Make-Map-List" data-toc-modified-id="Make-Map-List-2.1.2"><span class="toc-item-num">2.1.2&nbsp;&nbsp;</span>Make Map List</a></span></li></ul></li><li><span><a href="#Attribute-Value" data-toc-modified-id="Attribute-Value-2.2"><span class="toc-item-num">2.2&nbsp;&nbsp;</span>Attribute Value</a></span><ul class="toc-item"><li><span><a href="#Check-and-Clean-Tag" data-toc-modified-id="Check-and-Clean-Tag-2.2.1"><span class="toc-item-num">2.2.1&nbsp;&nbsp;</span>Check and Clean Tag</a></span></li><li><span><a href="#Make-Map-List" data-toc-modified-id="Make-Map-List-2.2.2"><span class="toc-item-num">2.2.2&nbsp;&nbsp;</span>Make Map List</a></span></li></ul></li></ul></li><li><span><a href="#Feature-Cleaning" data-toc-modified-id="Feature-Cleaning-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Feature Cleaning</a></span><ul class="toc-item"><li><span><a href="#Lemmatization" data-toc-modified-id="Lemmatization-3.1"><span class="toc-item-num">3.1&nbsp;&nbsp;</span>Lemmatization</a></span></li><li><span><a href="#Merge-Data" data-toc-modified-id="Merge-Data-3.2"><span class="toc-item-num">3.2&nbsp;&nbsp;</span>Merge Data</a></span></li></ul></li></ul></div>

# Loading Required libraries

import sys
import re
import pandas as pd
import pandas as pd
import nltk
import splitter
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import sent_tokenize
from sklearn.feature_extraction.text import CountVectorizer

# Tag Cleaning

data = pd.read_csv("usc_additional_tags USC.csv")

## Attribute Name

attribute_name = list(data.attribute_name.unique())

### Check and Clean Tag

cleaned_attribute_name_tag = [['material', 'clothing'],
 ['style'],
 ['primary', 'color'],
 ['upper', 'material'],
 ['wash'],
 ['trend'],
 ['pattern'],
 ['additional', 'color'],
 ['material'],
 ['material', 'purse'],
 ['strap', 'material'],
 ['print'],
 ['color'],
 ['belt', 'buckle', 'material'],
 ['belt', 'material'],
 ['sunglass', 'frame', 'material'],
 ['occasion'],
 ['category'],
 ['sizing'],
 ['length', 'top'],
 ['fit'],
 ['subcategory', 'top'],
 ['neckline'],
 ['gender'],
 ['embellishment'],
 ['sleeve', 'length'],
 ['dry', 'clean', 'only'],
 ['class', 'sneakers', 'and', 'athletic'],
 ['subcategory', 'shoe'],
 ['toe','style'],
 ['shoe', 'width'],
 ['toe', 'exposure'],
 ['closure', 'shoe'],
 ['length', 'coats', 'and', 'jackets'],
 ['class', 'blazers', 'coats', 'and', 'jackets'],
 ['closure', 'blazers', 'coats', 'and', 'jackets'],
 ['subcategory', 'blazers', 'coats', 'and', 'jackets'],
 ['class', 'dress'],
 ['length', 'one', 'piece'],
 ['subcategory', 'one', 'piece'],
 ['closure', 'one', 'piece'],
 ['length', 'pants', 'and', 'leggings'],
 ['closure', 'pants', 'and', 'leggings'],
 ['class', 'pants', 'and', 'leggings'],
 ['leg','style'],
 ['rise'],
 ['subcategory', 'bottom'],
 ['strap'],
 ['length', 'blazers'],
 ['length', 'skirts'],
 ['class', 'skirts'],
 ['closure', 'skirts'],
 ['closure', 'top'],
 ['subcategory', 'accessory'],
 ['closure', 'handbag'],
 ['class', 'handbags'],
 ['subcategory', 'sweater'],
 ['sheer'],
 ['subcategory', 'sweatshirt', 'and', 'hoodie'],
 ['shaft', 'height'],
 ['calf', 'width'],
 ['class', 'boots'],
 ['heels', 'hape'],
 ['heel', 'height'],
 ['length', 'blazers', 'coats', 'and', 'jackets'],
 ['class', 'jumpsuit', 'and', 'romper'],
 ['class', 'sandals'],
 ['length','shorts'],
 ['class', 'shorts'],
 ['rise', 'jeans'],
 ['leg','style', 'jeans'],
 ['length', 'jeans'],
 ['closure','shorts'],
 ['class', 'mules', 'and', 'slides'],
 ['class', 'booties'],
 ['class', 'pumps', 'and', 'heels'],
 ['closure','sweater'],
 ['class', 'flats'],
 ['class', 'sunglasses'],
 ['belt', 'buckles', 'hape'],
 ['belt', 'closure'],
 ['class', 'belts'],
 ['sweatshirt', 'and', 'hoodie', 'closure'],
 ['class', 'wedges'],
 ['class', 'slippers'],
 ['belt', 'width']]

attribute_name_map_list  = [['material'],
 ['style'],
 ['color'],
 ['material'],
 ['wash'],
 ['trend'],
 ['pattern'],
 ['color'],
 ['material'],
 ['material'],
 ['strap'],
 ['print'],
 ['color'],
 ['material'],
 ['material'],
 ['material'],
 ['occasion'],
 ['category'],
 ['sizing'],
 ['length'],
 ['fit'],
 ['category'],
 ['neckline'],
 ['gender'],
 ['embellishment'],
 ['length'],
 ['dry_clean_only'],
 ['class'],
 ['category'],
 ['toe_style'],
 ['width'],
 ['exposure'],
 ['closure'],
 ['length'],
 ['class'],
 ['closure'],
 ['category'],
 ['class',],
 ['length'],
 ['category'],
 ['closure'],
 ['length'],
 ['closure'],
 ['class'],
 ['leg_style'],
 ['rise'],
 ['category'],
 ['strap'],
 ['length'],
 ['length'],
 ['class'],
 ['closure'],
 ['closure'],
 ['category'],
 ['closure'],
 ['class'],
 ['category'],
 ['sheer'],
 ['category'],
 ['height'],
 ['width'],
 ['class'],
 ['shape'],
 ['height'],
 ['length'],
 ['class'],
 ['class'],
 ['length'],
 ['class'],
 ['rise'],
 ['leg_style'],
 ['length'],
 ['closure'],
 ['class'],
 ['class'],
 ['class'],
 ['closure'],
 ['class'],
 ['class'],
 ['shape'],
 ['closure'],
 ['class'],
 ['closure'],
 ['class'],
 ['class'],
 ['width']]

### Make Map List

attribute_name_map_dict = {}
for i, col_name in zip(range(len(attribute_name)),attribute_name):
    attribute_name_map_dict[col_name] = attribute_name_map_list[i][0]

## Attribute Value

attribute_value = list(data.attribute_value.unique())

### Check and Clean Tag

cleaned_attribute_value_tag = [['linen', 'blend'],
 ['cotton', 'blend'],
 ['modern'],
 ['business', 'casual'],
 ['classic'],
 ['blues'],
 ['leather'],
 ['glam'],
 ['edgy'],
 ['blacks'],
 ['medium', 'blue'],
 ['denim'],
 ['casual'],
 ['raw', 'hem'],
 ['pure', 'cotton'],
 ['polyester'],
 ['yellows'],
 ['viscose'],
 ['retro'],
 ['oranges'],
 ['white'],
 ['whites'],
 ['abstract'],
 ['multi'],
 ['boho'],
 ['pure', 'silk'],
 ['navy'],
 ['nylon'],
 ['dark', 'browns'],
 ['grays'],
 ['dark', 'blue'],
 ['clean'],
 ['synthetic'],
 ['androgynous'],
 ['greens'],
 ['pinks'],
 ['floral'],
 ['pure', 'wool'],
 ['stripe'],
 ['wool', 'blend'],
 ['cashmere', 'blend'],
 ['pure', 'cashmere'],
 ['romantic'],
 ['beige'],
 ['golds'],
 ['silvers'],
 ['alpaca'],
 ['suede'],
 ['stripe', 'vertical'],
 ['athleisure'],
 ['dots'],
 ['satin', 'charmeuse'],
 ['geometric'],
 ['animal'],
 ['raffia'],
 ['silk', 'blend'],
 ['silk'],
 ['elastane'],
 ['pure', 'linen'],
 ['color', 'block'],
 ['spandex'],
 ['light', 'brown'],
 ['ramie'],
 ['black'],
 ['canvas'],
 ['velvet'],
 ['reds'],
 ['light', 'brown'],
 ['shearling'],
 ['rayon'],
 ['acetate'],
 ['stripe', 'horizontal'],
 ['graphic'],
 ['burgundies'],
 ['plaid'],
 ['snakeskin'],
 ['modal'],
 ['light', 'blue'],
 ['lyocell'],
 ['tiedye'],
 ['purples'],
 ['metal'],
 ['acrylic'],
 ['cupro'],
 ['mohair'],
 ['multi', 'print'],
 ['camouflage'],
 ['distressed'],
 ['calf', 'hair'],
 ['teal'],
 ['tropical'],
 ['faux', 'leather'],
 ['logo'],
 ['none'],
 ['faux', 'fur'],
 ['Grey'],
 ['chenille'],
 ['corduroy'],
 ['houndstooth'],
 ['monogram'],
 ['paisley'],
 ['trim'],
 ['linen'],
 ['pony', 'hair'],
 ['faux', 'shearling'],
 ['crepe', 'de','chine'],
 ['fleece'],
 ['pinstripe'],
 ['denim'],
 ['acid'],
 ['coated'],
 ['chambray'],
 ['sateen'],
 ['gold'],
 ['raw'],
 ['beaded'],
 ['tech'],
 ['patches'],
 ['straw'],
 ['tweed'],
 ['chiffon'],
 ['wool'],
 ['silver'],
 ['brass'],
 ['twill'],
 ['patent', 'leather'],
 ['plastic'],
 ['weekend'],
 ['top'],
 ['regular'],
 ['mid', 'length', 'at', 'hips'],
 ['straight', 'regular'],
 ['tee'],
 ['crew', 'neck'],
 ['women'],
 ['embroidery'],
 ['short','sleeve'],
 ['no'],
 ['long','sleeve'],
 ['workout'],
 ['semifitted'],
 ['active'],
 ['day', 'tonight'],
 ['pull', 'on'],
 ['sneakers', 'athletic'],
 ['round', 'toe'],
 ['medium'],
 ['closed', 'toe'],
 ['shoe'],
 ['zipper'],
 ['lace', 'top'],
 ['laces'],
 ['low', 'top'],
 ['training'],
 ['yes'],
 ['relaxed'],
 ['bomber', 'varsity'],
 ['jacket'],
 ['cold', 'weather'],
 ['blazers,', 'coats,', 'jackets'],
 ['anorak,', 'raincoat,', 'windbreaker'],
 ['mesh'],
 ['parka', 'puffer'],
 ['maxi', 'at', 'and', 'below', 'calf'],
 ['oversized'],
 ['casual', 'dress'],
 ['maxi', 'past', 'calf', 'and', 'below'],
 ['one', 'piece'],
 ['maxi'],
 ['dress'],
 ['cropped', 'above', 'waistline'],
 ['short', 'above', 'the', 'knee'],
 ['plunge', 'neck'],
 ['belt'],
 ['Aline'],
 ['ruffles'],
 ['cropped'],
 ['bottom'],
 ['zip', 'fly', 'with', 'button'],
 ['cargo'],
 ['straight'],
 ['mid', '89'],
 ['pants,', 'leggings'],
 ['studs'],
 ['drop', 'waist'],
 ['vacation'],
 ['halter'],
 ['adjustable'],
 ['single', 'breasted'],
 ['blazer'],
 ['buttoned', 'front'],
 ['night', 'out'],
 ['sheath'],
 ['knee','length'],
 ['fitted', 'tailored'],
 ['vneck'],
 ['work'],
 ['midi', 'mid', 'calf'],
 ['back', 'zip'],
 ['skirts'],
 ['collared'],
 ['crop', 'top'],
 ['button', 'down'],
 ['tiered'],
 ['side', 'zip'],
 ['handbags'],
 ['snap'],
 ['accessory'],
 ['cross', 'body', 'bags'],
 ['magnetic'],
 ['shoulder', 'bags'],
 ['tunic'],
 ['sweater'],
 ['short', 'at', 'waistline'],
 ['mock', 'neck'],
 ['sleeveless'],
 ['tote', 'bags'],
 ['open'],
 ['split', 'neck'],
 ['square', 'neck'],
 ['pullover'],
 ['turtleneck'],
 ['semisheer'],
 ['zip', 'up'],
 ['slit'],
 ['bucket', 'bags'],
 ['drawstring'],
 ['long', 'below', 'hips'],
 ['high', 'neck'],
 ['hooded'],
 ['sweatshirt,', 'hoodie'],
 ['knee', 'high'],
 ['slouch'],
 ['boots'],
 ['cone'],
 ['pointed', 'toe'],
 ['low'],
 ['block'],
 ['bootie'],
 ['ankle'],
 ['skinny'],
 ['denim', 'jeans'],
 ['high', 'over', '9'],
 ['mid', 'calf'],
 ['camisole'],
 ['front', 'zip'],
 ['trousers'],
 ['pumps', 'heels'],
 ['stiletto'],
 ['shirt', 'dress'],
 ['double', 'breasted'],
 ['buttoned', 'back'],
 ['jumpsuit'],
 ['open', 'toe'],
 ['ankles', 'trap'],
 ['sandals'],
 ['lace', 'up'],
 ['wrap'],
 ['wide', 'leg'],
 ['zip', 'fly', 'with', 'hook'],
 ['asymmetrical'],
 ['sweetheart'],
 ['short','35'],
 ['shorts'],
 ['sheer'],
 ['casual','shorts'],
 ['slide'],
 ['wedge'],
 ['tieback'],
 ['fringe'],
 ['blouse'],
 ['baggy'],
 ['flat'],
 ['platform', 'flatform'],
 ['slits'],
 ['slip', 'dress'],
 ['pussy', 'bow'],
 ['back', 'zipper'],
 ['overcoat'],
 ['coat'],
 ['5','pocket','pant','non','denim'],
 ['boot', 'cut'],
 ['flare'],
 ['slim'],
 ['designer'],
 ['scoop', 'neck'],
 ['moto'],
 ['jewel'],
 ['backpacks'],
 ['sleeve'],
 ['belt bags,', 'fanny pack'],
 ['cold', 'shoulder'],
 ['military', 'utility'],
 ['tie', 'front'],
 ['puff', 'sleeve'],
 ['high'],
 ['buckle'],
 ['mini','mid', 'thigh', 'and', 'above'],
 ['knit'],
 ['chino', 'khaki'],
 ['button', 'front'],
 ['quilted'],
 ['front', 'zip', 'with', 'button'],
 ['long','10'],
 ['mules', 'slides'],
 ['mule'],
 ['cowl', 'neck'],
 ['clutches', 'pouches'],
 ['trapeze','swing'],
 ['slip'],
 ['strappy'],
 ['hook', 'and', 'loop'],
 ['espadrille'],
 ['combat'],
 ['tie', 'neck'],
 ['kitten'],
 ['thigh', 'high'],
 ['boat', 'neck'],
 ['clasp'],
 ['off', 'shoulder'],
 ['pencil'],
 ['flats'],
 ['belted'],
 ['trench'],
 ['capri'],
 ['low', 'under', '8'],
 ['dorsay'],
 ['short'],
 ['shift'],
 ['open', 'front'],
 ['cardigan'],
 ['sweater', 'dress'],
 ['joggers,', 'sweatpants'],
 ['strapless'],
 ['Bermuda'],
 ['tie'],
 ['booties'],
 ['peplum'],
 ['leggings'],
 ['high', 'top'],
 ['dad', 'sneaker', 'platform'],
 ['scarves,', 'wraps'],
 ['slip', 'on'],
 ['funnel'],
 ['shell'],
 ['body', 'con'],
 ['peep', 'toe'],
 ['blazer', 'dress'],
 ['sequins'],
 ['tank'],
 ['keyhole'],
 ['shirt', 'jacket'],
 ['western', 'cowboy'],
 ['ballet'],
 ['slingback'],
 ['trucker'],
 ['cateye'],
 ['sunglasses'],
 ['surplice'],
 ['cap','sleeve'],
 ['Chelsea'],
 ['rain'],
 ['tshirt', 'dress'],
 ['business', 'casual', 'dress'],
 ['flap'],
 ['cape,', 'poncho,', 'wrap'],
 ['lace'],
 ['spaghetti'],
 ['sundress'],
 ['bodysuit'],
 ['tall'],
 ['pea', 'coat'],
 ['romper'],
 ['buckles'],
 ['Mary', 'Jane'],
 ['elastic','stretch'],
 ['vest'],
 ['pleated'],
 ['empire', 'waist'],
 ['laptops,', 'briefcases'],
 ['halfway', 'buttoned'],
 ['snaps'],
 ['slingbacks'],
 ['maternity'],
 ['henley'],
 ['lock'],
 ['paper', 'bag', 'waist'],
 ['pool','slide'],
 ['high', 'low'],
 ['boyfriend'],
 ['mandarin', 'collar'],
 ['crystals'],
 ['belts'],
 ['square', 'toe'],
 ['hobo', 'bags'],
 ['flip', 'flop'],
 ['button', 'up'],
 ['Hook','eye'],
 ['teddy'],
 ['14','zip'],
 ['halfway', 'zipper'],
 ['back', 'tie'],
 ['shawl', 'collar'],
 ['sports', 'bra'],
 ['plus'],
 ['active', 'track'],
 ['culotte'],
 ['full'],
 ['very', 'short', '3'],
 ['polo'],
 ['poncho'],
 ['track', 'pants'],
 ['bike'],
 ['sweater', 'coat'],
 ['jersey', 'knit'],
 ['rhinestone'],
 ['wide'],
 ['backless'],
 ['narrow'],
 ['raised', 'sole'],
 ['wedges'],
 ['motorcycle'],
 ['front', 'zip', 'with', 'hook'],
 ['cutoffs'],
 ['duster'],
 ['mid','510'],
 ['oxford'],
 ['loafer'],
 ['skort'],
 ['oval'],
 ['double', 'prong'],
 ['cheeky'],
 ['strap'],
 ['satchels'],
 ['wallets', 'card', 'cases'],
 ['rectangular'],
 ['Velcro'],
 ['inward', 'curve'],
 ['band', 'collar'],
 ['epaulets'],
 ['slippers'],
 ['moccasin'],
 ['beach', 'bags'],
 ['rectangle'],
 ['aviator','square'],
 ['one', 'shoulder'],
 ['front', 'tie'],
 ['twisted'],
 ['bustier'],
 ['over','the', 'knee'],
 ['aviator', 'teardrop'],
 ['sock'],
 ['skinny', 'up', 'to', '1'],
 ['hook'],
 ['tennis'],
 ['draped']]

joined_attribute_value_tag = []
for i in cleaned_attribute_value_tag:
    joined_attribute_value_tag.append(' '.join(i))

### Make Map List

attribute_value_map_dict = {}
for i, col_name in zip(range(len(attribute_value)),attribute_value):
    attribute_value_map_dict[col_name] = joined_attribute_value_tag[i]

data_cleaned= data[['product_id','product_color_id']]
data_cleaned['attribute_name'] = data.attribute_name.map(attribute_name_map_dict)
data_cleaned['attribute_value'] = data.attribute_value.map(attribute_value_map_dict)

data_cleaned.to_csv('cleaned_usc_additional_tags USC.csv', index=False)

# Feature Cleaning

clean_data = data_cleaned.copy()
clean_data = clean_data.drop(columns = 'product_color_id')
clean_data = clean_data.drop_duplicates()

## Lemmatization

lemmatizer = WordNetLemmatizer()

for i in range(len(clean_data)):
    review = clean_data.iloc[i,1].split()
    new_words = []
    for word in review:
        new_words.append(lemmatizer.lemmatize(word))
    cleaned_review = " ".join(new_words)
    clean_data.iloc[i,1] = cleaned_review

lemmatizer = WordNetLemmatizer()

for i in range(len(clean_data)):
    review = clean_data.iloc[i,2].split()
    new_words = []
    for word in review:
        new_words.append(lemmatizer.lemmatize(word))
    cleaned_review = " ".join(new_words)
    clean_data.iloc[i,2] = cleaned_review

## Merge Data

merge = clean_data.groupby(['product_id','attribute_name'])['attribute_value'].apply(lambda x: ','.join(x)).reset_index()

merge2 = merge.drop(columns = 'product_id')
merge2 = merge2.drop_duplicates()
attr_name = merge2.groupby(['attribute_name'])['attribute_value'].apply(lambda x: ','.join(x)).reset_index()
value_name = attr_name.attribute_value.apply(lambda x: x.split(',')).apply(lambda x: set(x))

final = merge.pivot(index='product_id', columns='attribute_name', values='attribute_value').reset_index().rename_axis(None, axis=1)

attr_name.to_csv('attribute_name2.csv')
final.to_csv('cleaned_tags_for_future_use.csv')
