==> Project Folder/2. Create Tags and Features.ipynb <==
<h1>Create Tags/Features<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Loading-Libraries" data-toc-modified-id="Loading-Libraries-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Loading Libraries</a></span></li><li><span><a href="#Define-Functions" data-toc-modified-id="Define-Functions-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Define Functions</a></span></li><li><span><a href="#Load-Data" data-toc-modified-id="Load-Data-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Load Data</a></span></li><li><span><a href="#Text-Preprocessing-(combine,-clean-and-lemmatize-data/columns)" data-toc-modified-id="Text-Preprocessing-(combine,-clean-and-lemmatize-data/columns)-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Text Preprocessing (combine, clean and lemmatize data/columns)</a></span><ul class="toc-item"><li><span><a href="#Check-Dataset" data-toc-modified-id="Check-Dataset-4.1"><span class="toc-item-num">4.1&nbsp;&nbsp;</span>Check Dataset</a></span></li><li><span><a href="#Combine-Product-and-Additional_tags-(look-for-patterns-for-attribute_category)" data-toc-modified-id="Combine-Product-and-Additional_tags-(look-for-patterns-for-attribute_category)-4.2"><span class="toc-item-num">4.2&nbsp;&nbsp;</span>Combine Product and Additional_tags (look for patterns for attribute_category)</a></span></li><li><span><a href="#Combine-Columns-for-Extracting-Tags" data-toc-modified-id="Combine-Columns-for-Extracting-Tags-4.3"><span class="toc-item-num">4.3&nbsp;&nbsp;</span>Combine Columns for Extracting Tags</a></span><ul class="toc-item"><li><span><a href="#For-category-(first-priority)" data-toc-modified-id="For-category-(first-priority)-4.3.1"><span class="toc-item-num">4.3.1&nbsp;&nbsp;</span>For category (first priority)</a></span><ul class="toc-item"><li><span><a href="#Combine-brand_category-and-name-columns" data-toc-modified-id="Combine-brand_category-and-name-columns-4.3.1.1"><span class="toc-item-num">4.3.1.1&nbsp;&nbsp;</span>Combine <code>brand_category</code> and <code>name</code> columns</a></span></li><li><span><a href="#Clean-the-combined-column" data-toc-modified-id="Clean-the-combined-column-4.3.1.2"><span class="toc-item-num">4.3.1.2&nbsp;&nbsp;</span>Clean the combined column</a></span></li></ul></li><li><span><a href="#For-catgeory-(second-priority)-and-other-tags" data-toc-modified-id="For-catgeory-(second-priority)-and-other-tags-4.3.2"><span class="toc-item-num">4.3.2&nbsp;&nbsp;</span>For catgeory (second priority) and other tags</a></span><ul class="toc-item"><li><span><a href="#Make-location-tokens" data-toc-modified-id="Make-location-tokens-4.3.2.1"><span class="toc-item-num">4.3.2.1&nbsp;&nbsp;</span>Make location tokens</a></span></li><li><span><a href="#Combine-brand_category,-name-,-details-,description,-brand_name-columns,-and-clean-the-combined-column" data-toc-modified-id="Combine-brand_category,-name-,-details-,description,-brand_name-columns,-and-clean-the-combined-column-4.3.2.2"><span class="toc-item-num">4.3.2.2&nbsp;&nbsp;</span>Combine <code>brand_category</code>, <code>name</code> , <code>details</code> ,<code>description</code>, <code>brand_name</code> columns, and clean the combined column</a></span></li></ul></li></ul></li></ul></li><li><span><a href="#Major-Feature-Processing" data-toc-modified-id="Major-Feature-Processing-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Major Feature Processing</a></span><ul class="toc-item"><li><span><a href="#Category" data-toc-modified-id="Category-5.1"><span class="toc-item-num">5.1&nbsp;&nbsp;</span>Category</a></span><ul class="toc-item"><li><span><a href="#Define-Pattern-for-Each-Category" data-toc-modified-id="Define-Pattern-for-Each-Category-5.1.1"><span class="toc-item-num">5.1.1&nbsp;&nbsp;</span>Define Pattern for Each Category</a></span></li><li><span><a href="#Find-Detailed-Category---One/Multiple-for-Each-Product" data-toc-modified-id="Find-Detailed-Category---One/Multiple-for-Each-Product-5.1.2"><span class="toc-item-num">5.1.2&nbsp;&nbsp;</span>Find Detailed Category - One/Multiple for Each Product</a></span></li><li><span><a href="#Find-General-Category---Only-One-for-Each-Product" data-toc-modified-id="Find-General-Category---Only-One-for-Each-Product-5.1.3"><span class="toc-item-num">5.1.3&nbsp;&nbsp;</span>Find General Category - Only One for Each Product</a></span></li></ul></li><li><span><a href="#Gender" data-toc-modified-id="Gender-5.2"><span class="toc-item-num">5.2&nbsp;&nbsp;</span>Gender</a></span></li><li><span><a href="#Season" data-toc-modified-id="Season-5.3"><span class="toc-item-num">5.3&nbsp;&nbsp;</span>Season</a></span></li><li><span><a href="#Other-Tags" data-toc-modified-id="Other-Tags-5.4"><span class="toc-item-num">5.4&nbsp;&nbsp;</span>Other Tags</a></span></li><li><span><a href="#Create-a-column-for-all-the-features" data-toc-modified-id="Create-a-column-for-all-the-features-5.5"><span class="toc-item-num">5.5&nbsp;&nbsp;</span>Create a column for all the features</a></span></li></ul></li><li><span><a href="#Final-data" data-toc-modified-id="Final-data-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>Final data</a></span></li></ul></div>

We used the above steps to created tags and features. **Please run this notebook to create tags and features**. Those tags will be helpful for our model and recommendation function.

# Loading Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import re
import collections
from collections import Counter
import nltk
from nltk.tokenize import sent_tokenize, word_tokenize
from nltk.stem import WordNetLemmatizer
from nltk.stem import PorterStemmer 
from nltk.corpus import stopwords, wordnet
from IPython.display import Markdown, display
import copy
import nltk
#nltk.download('averaged_perceptron_tagger')

# Define Functions

# reference: https://gist.github.com/gaurav5430/9fce93759eb2f6b1697883c3782f30de#file-nltk-lemmatize-sentences-py
lemmatizer = WordNetLemmatizer()

# function to convert nltk tag to wordnet tag
def nltk_tag_to_wordnet_tag(nltk_tag):
    if nltk_tag.startswith('J'):
        return wordnet.ADJ
    elif nltk_tag.startswith('V'):
        return wordnet.VERB
    elif nltk_tag.startswith('N'):
        return wordnet.NOUN
    elif nltk_tag.startswith('R'):
        return wordnet.ADV
    else:
        return None

def lemmatize_sentence(sentence):
    # tokenize the sentence and find the POS tag for each token
    nltk_tagged = nltk.pos_tag(word_tokenize(sentence))
    # tuple of (token, wordnet_tag)
    wordnet_tagged = map(lambda x: (
        x[0], nltk_tag_to_wordnet_tag(x[1])), nltk_tagged)
    lemmatized_sentence = []
    for word, tag in wordnet_tagged:
        if tag is None:
            # if there is no available tag, append the token as is
            lemmatized_sentence.append(word)
        else:
            # else use the tag to lemmatize the token
            lemmatized_sentence.append(lemmatizer.lemmatize(word, tag))
    return ' '.join(lemmatized_sentence)

# reference https://www.codenong.com/2277352/
from collections import Counter

#split words by uppercase, for example, flatShoe --> flat Shoe
def split_on_uppercase(s, keep_contiguous=False):
    string_length = len(s)
    is_lower_around = (lambda: s[i-1].islower() or
                       string_length > (i + 1) and s[i + 1].islower())
    start = 0
    parts = []
    for i in range(1, string_length):
        if s[i].isupper() and (not keep_contiguous or is_lower_around()):
            parts.append(s[start: i])
            start = i
    parts.append(s[start:])

    return ' '.join(parts)

# count words frequency
def count_words(lines, delimiter=" "):
    words = Counter()

    for line in lines:
        for word in line.split(delimiter):
            words[word] += 1 

    return words

# Load Data

product_data = pd.read_excel('Behold+product+data+04262021.xlsx')
tags_data = pd.read_csv('usc_additional_tags USC.csv')

# Text Preprocessing (combine, clean and lemmatize data/columns)

## Check Dataset

Drop useless columns in product dataset

product_drop=['brand_canonical_url','created_at','brand_description','product_active']

product_data_dropped=product_data.drop(columns=product_drop)
product = copy.deepcopy(product_data_dropped)

Check if there is a duplicated product

product.duplicated('product_id').sum()

Check different types of tags

tags_data['attribute_name'].unique()

Create dataframe for attribute_category

category = tags_data[tags_data.attribute_name == 'category'].drop(
    columns=['product_color_id', "attribute_name"]).rename(columns={"attribute_value": "attribute_category"})
category = category.drop_duplicates()

category.head()

## Combine Product and Additional_tags (look for patterns for attribute_category)

This part is only for checking patterns for each category.

- Create a combined DataFrame by merging category data (from additional_tags dataset) and product dataset.
- We can use this combined DataFrame to compare product information and categories, and find the patterns for each attribute_category.

combined=pd.merge(product, category, how='left', on='product_id')

# find duplicated product after combination
combined.duplicated('product_id').sum()

# combine attribute_category to remove duplicated products
combined=combined.drop([58047])
combined.loc[57814,'attribute_category']='top sweatshirt'
combined=combined.drop([57815])
combined=combined.drop([59160])
combined.loc[59171,'attribute_category']='top coat'
combined=combined.drop([59172])
combined=combined.drop([59176])
combined=combined.drop([60296])
combined.loc[60299,'attribute_category']='jacket sweater'
combined=combined.drop([60300])

#check the category from tag document
combined.attribute_category.value_counts()

## Combine Columns for Extracting Tags

Add columns to product dataset by combining original columns that are useful to extract patterns. We will only use the new columns to extract tags.

### For category (first priority)

 - We will extract "category"  first since we need this tag in our recommendation function.
 - We thought `brand_category` and `name` columns would have the most accurate information to extract category. We will tag category by using the two columns first and for the rest of the data that cannot find category from those two columns, we will use columns such as `details`, `description` to do the extraction.

#### Combine `brand_category` and `name` columns

product['choose_category_1']=product['brand_category'].map(str)+' '+product['name'].map(str)

#### Clean the combined column

for col in ['choose_category_1']:
    product[col] = product[col].str.replace('\n', ' ')
    product[col] = product[col].str.replace('\r', ' ')
    product[col] = product[col].str.replace(r'\\', '', regex = True)
    product[col] = product[col].str.replace(r'[^\w\s]|_', ' ', regex = True)
    product[col] = product[col].str.replace(r'[0-9]', '', regex = True)
product['choose_category_1'].tail()

product['choose_category_1'] = product['choose_category_1'].apply(str).apply(lambda x: split_on_uppercase(x, keep_contiguous=True))
product['choose_category_1'].tail()

#lemmatize choose_category_1 column
product['lemm_choose_category_1'] = product['choose_category_1'].str.lower().apply(str).apply(lambda x: lemmatize_sentence(x))

### For catgeory (second priority) and other tags

Combine `brand_category`, `name` , `details` ,`description`, `brand_name` columns. The new column will be used to further extract category and all other tags.

#### Make location tokens

# location combos

location_tokens = {}
location_tokens['new york(?: city)?'] = 'ny'
location_tokens['nyc'] = 'ny'
location_tokens['the usa'] = 'usa'
location_tokens['the united states'] = 'usa'
location_tokens['united states(?: of america)?'] = 'usa'
location_tokens['los angeles'] = 'la'
location_tokens['north carolina'] = 'northcarolina'
location_tokens['south carolina'] = 'southcarolina'
location_tokens['new jersey'] = 'newjersey'
location_tokens['rhode island'] = 'rhodeisland'
location_tokens['new hampshire'] = 'newhampshire'
location_tokens['west virginia'] = 'westvirginia'
location_tokens['north dakota'] = 'northdakota'

location_tokens['hong kong'] = 'hongkong'
location_tokens['guang zhou'] = 'guangzhou'

#### Combine `brand_category`, `name` , `details` ,`description`, `brand_name` columns, and clean the combined column

product['total'] = product['brand_category'].map(str) + ' ' + product['name'].map(str) + ' ' + product['details'].map(str) + ' ' + product['description'].map(str) + ' ' + product['brand_name'].map(str)

# clean the total column
for col in ['total']:
    product[col] = product[col].str.replace('\n', ' ')
    product[col] = product[col].str.replace('\r', ' ')
    product[col] = product[col].str.replace(r'\\', '', regex=True)
    product[col] = product[col].str.replace(r'[^\w\s\%]|_', ' ', regex=True)

product['total'] = product['total'].apply(str).apply(lambda x: split_on_uppercase(
    x, keep_contiguous=True)).str.lower().str.replace('nan', ' ').str.replace('  ', ' ').str.replace('  ', ' ')

for key in location_tokens:
    product['total'] = product['total'].str.replace(
        rf'\b({key})\b', f'{location_tokens[key]}', regex=True)
product['total'].tail()

#lemmatize total
product['lemm_total'] = product['total'].str.lower().apply(str).apply(lambda x: lemmatize_sentence(x))

product['lemm_total'].tail()

# Major Feature Processing

## Category

### Define Pattern for Each Category
- only use simple words since column is cleaned/lemmatized

top_re=r'\b(top|shirt|tee|blouse|tank|tshirt|knitwear|vest|camisole|overshirt|long sleeve|v neck|tunic|turtleneck|shell|sleeve|polo|cami|slvls|vneck|shrt|pckt t|crew|sleeveless|sweatshirt|longsleeve|hoodie|pullover|sweat|crewneck|pullovr|swtshrt|sweater|cardigan|knitwear|swtr)\b'
outwear_re = r'\b(jacket|coat|blazer|outerwear|trench|parka|anorak|puffer|overcoat|topcoat|bomber|trucker|jckt|blazr|jkt)\b'
bottom_re = r'\b(jean|pant|jogger|skirt|short|trouser|legging|leg|skort|bottom|low straight|loose straight|straight fit|loose fit|relax straight|relaxed straight|vintage straight|slim taper|sweatpants?|miniskirt|culotte|bootcut|dojo|crop|skinny|stretch denim|slimmy|high rise|low rise|mid rise|garment|paxtyn|ankle|denim slimmy|high waist|flare|skny|exaggerate kick)\b'
shoe_re = r'\b(sandal|booty|shoe|sneaker|boot|pump|flat|heel|bootie|mule|espadrille|oxford|loafer|slipper|flip flop|clog|derby|slip on|wedge|slide|nappa|leather upper)\b'
onepiece_re = r'\b(dress|jumpsuit|shirtdress|dressesandjumpsuits?|bodysuit|housedress|rompers?|kimono|robe|one piece|onepiece|minidress|denim set|pj set|short set|lounge set|sleep set|cami set|trouser set|pajama set|pc set|bra set|slipdress|pantsuit|shirtdress|raincoat|caftan|suit|slip|jump|jumpsut)\b'
accessory_re = r'\b(bags?|tote|crossbody|handbags?|belt|tie|accessory|ladybag|backpack|gymsack|cardholder|glove|wristlet|sunglass|satchel|leash|laptop|camera|jewelry|wallet|eyewear|scarf|scarves|clutch|poncho|calendar|bangle|keychain|mat|postcard|tumbler|pouch|hobo|hat|earring|barrette|haircalf|box|cape|mask|mitt|sock|wrap|jasper|bracelet|necklace|scrunchie|biker|millinery|card|fedora|felt|pack|napkin|shawl|hook|chain|spray|diffuser|brim|cap|blanket|josefina|bottle|basket|eye|extérieure|paint|mousse|glass|collection|acrylic|accessory|clip|beanie|beret)\b'
underwear_re = r'\b(bra|underwear|bralett|bralette|thong|panty|underwire|brakini)\b'
othercloth_re=r'\b(wears?|classic fit|regular fit|denim|cloth|othercloth)\b'
home_re=r'\b(candle|vase|mug|pot|chair|mist|towel|lamp|glasswear|plate|pillow|table|coverlet|sofa|bed|desk|bowl|dish|bath|jar|soap|handwoven|chèvre|wash|home)\b'
cosmetics_re=r'\b(sanitizer|cream|lip|oil|perfume|perfum|nail)\b'
book_re=r'\b(books?|lookbook|workbook)\b'
archive_re=r'\barchive\b'

### Find Detailed Category - One/Multiple for Each Product

- Extract detailed category such as blouse,tank and coat. Each product may have more than one detailed categories.

def detialedCategory_tag(txt):
    val = []
    txt = str(txt)
    if re.findall(shoe_re, txt, re.IGNORECASE ):
        val = re.findall(shoe_re, txt, re.IGNORECASE )
    elif re.findall(onepiece_re, txt, re.IGNORECASE ):
        val = re.findall(onepiece_re, txt, re.IGNORECASE )
    elif re.findall(bottom_re, txt, re.IGNORECASE ):
        val = re.findall(bottom_re, txt, re.IGNORECASE )
    elif re.findall(top_re, txt, re.IGNORECASE ):
        val = re.findall(top_re, txt, re.IGNORECASE )
    elif re.findall(outwear_re, txt, re.IGNORECASE ):
        val = re.findall(outwear_re, txt, re.IGNORECASE )
    elif re.findall(underwear_re, txt, re.IGNORECASE ):
        val = re.findall(underwear_re, txt, re.IGNORECASE )
    elif re.findall(accessory_re, txt, re.IGNORECASE ):
        val = re.findall(accessory_re, txt, re.IGNORECASE )
    #find other clothing after search above categories
    elif re.findall(othercloth_re, txt, re.IGNORECASE ):
        val = ['othercloth']
    elif re.findall(home_re, txt, re.IGNORECASE ):
        val = re.findall(home_re, txt, re.IGNORECASE )
    elif re.findall(cosmetics_re, txt, re.IGNORECASE ):
        val = re.findall(cosmetics_re, txt, re.IGNORECASE )
    elif re.findall(book_re, txt, re.IGNORECASE ):
        val = re.findall(book_re, txt, re.IGNORECASE )
    elif re.findall(archive_re, txt, re.IGNORECASE ):
        val = re.findall(archive_re, txt, re.IGNORECASE )
    else:
        val=['unknow']
    return ' '.join(val)

# first priority - more accurate to find category
product['detailed_category'] = product['lemm_choose_category_1'].apply(lambda x: detialedCategory_tag(x))
# second priority - less accurate but still usefull
for i in product[product['detailed_category'] == 'unknow'].index:
    product.loc[i, 'detailed_category'] = detialedCategory_tag(product.loc[i, 'lemm_total'])

#remove duplicated words in detailed_category column
product["detailed_category"]=product["detailed_category"].str.split(" ").map(collections.OrderedDict.fromkeys).str.join(" ")

### Find General Category - Only One for Each Product

- Extract general category such as top, bottom and onepiece. Each product only has one general category.

def generalCategory(txt):  
    txt = str(txt)
    val = np.nan
    if re.search(shoe_re, txt, re.IGNORECASE ):
        val = "shoe"
    elif re.search(onepiece_re, txt, re.IGNORECASE ):
        val = "onepiece"    
    elif re.search(bottom_re, txt, re.IGNORECASE ):
        val = "bottom"
    elif re.search(top_re, txt, re.IGNORECASE ):
        val = "top"
    elif re.search(outwear_re, txt, re.IGNORECASE ):
        val = "outwear"
    elif re.search(underwear_re, txt, re.IGNORECASE ):
        val = "underwear"
    elif re.findall(othercloth_re, txt, re.IGNORECASE ):
        val = "othercloth"
    elif re.search(accessory_re, txt, re.IGNORECASE ):
        val = "accessory"
    elif re.search(home_re, txt, re.IGNORECASE ):
        val = "home"
    elif re.search(cosmetics_re, txt, re.IGNORECASE ):
        val = "cosmetics"
    elif re.search(book_re, txt, re.IGNORECASE ):
        val = "book"
    elif re.search(archive_re, txt, re.IGNORECASE ):
        val = "archive"
    else:
        val = "unknow"
    return val

product['general_category']=product['detailed_category'].apply(lambda x: generalCategory(x))

## Gender

- Extract gender such as man, woman, boy, girl and baby. Each product only has one gender.

# define the pattern of gender
man_pattern=r'\b(?:grand(?:pas?|fathers?)|sons?|males?|fathers?|dads?|dadd(?:y|ies)|pa(?:pas?)?|man|men|husbands?|uncles?|brothers?|bros?|nephews?|boyfriends?)\b'
woman_pattern=r'\b(?:grand(?:mas?|mothers?)|daughters?|feminines?|females?|mothers?|woman|women|moms?|momm(?:y|ies)|ma(?:mas?)?|wi(?:fe|ves)|aunts?|aunt(?:y|ies?)|sisters?|nieces?|girlfriends?|skirts?|dress|minidress|miniskirts?)\b'
boy_pattern=r'\b(boys?|grandsons?)\b'
girl_pattern=r'\b(granddaughters?|girls?)\b'
baby_pattern=r'\b(baby|babies)\b'

def Gender(txt):
    txt = str(txt)
    val = np.nan
    if re.search(man_pattern, txt, re.IGNORECASE ):
        val = 'man'
    elif re.search(woman_pattern, txt, re.IGNORECASE ):
        val = 'woman'
    elif re.search(boy_pattern, txt, re.IGNORECASE ):
        val = 'boy'
    elif re.search(girl_pattern, txt, re.IGNORECASE ):
        val = 'girl'
    elif re.search(baby_pattern, txt, re.IGNORECASE ):
        val = 'baby'
    return val

product['gender']=product['lemm_total'].apply(lambda x: Gender(x))

product['gender'].notnull().sum()

## Season

- Extract season such as spring, summer, fall, and winter. Each product only has one season.

# define the pattern of the season
spring_pattern=r'\b(springs?|springtimes?|march|april|may)\b'
summer_pattern=r'\b(summers?|summertimes?|june|july|august)\b'
fall_pattern=r'\b(falls?|falltimes?|autumns?|september|october|november)\b'
winter_pattern=r'\b(winters?|wintertimes?|december|january|february)\b'

def Season(txt):
    txt = str(txt)
    val = np.nan
    if re.search(spring_pattern, txt, re.IGNORECASE ):
        val = 'spring'
    elif re.search(summer_pattern, txt, re.IGNORECASE ):
        val = 'summer'
    elif re.search(fall_pattern, txt, re.IGNORECASE ):
        val = 'fall'
    elif re.search(winter_pattern, txt, re.IGNORECASE ):
        val = 'winter'
    return val

product['season']=product['lemm_total'].apply(lambda x: Season(x))

product['season'].notnull().sum()

## Other Tags

- Extract more tags such locations, colors, sizing, style etc. Each product may have more than one values for the same tag.

# define patterns for each tags
# these regex patterns come from attribute_name2.csv
attributes_regex = {}
attributes_regex['class'] = 'fanny packs?|business casuals?|ballets?|trench(?:es)?|vests?|ponchos?|double prongs?|backpacks?|booty|booties|denims?|western cowboy|mules?|5 pocket pant non denim|flip flops?|slingbacks?|slits?|hooded|chelsea|dorsay|backless|beach bags?|loafers?|high tops?|wraps?|paper bag waists?|hobo bags?|skorts?|parka puffers?|slides?|cheeky|tshirts?|dress(?:es)?|bikes?|high low|sundress|track pants?|overcoats?|lace up|bomber varsity|casual dress(?:es)?|distressed|drop waists?|low tops?|bucket bags?|oxfords?|tiered|ankle trap|empire waists?|tennis|training|pleated|pullovers?|teddy|quilted|body cons?|cateye|designers?|peplums?|motos?|laptops?|leggings?|cargos?|pool slides?|asymmetrical|briefcases?|trousers?|motorcycles?|active|belt bags?|trapeze swings?|slip on|military utility|windbreakers?|peep toes?|shirts?|jackets?|pencil|oval|sheaths?|business casual dress(?:es)?|pea coats?|bermuda|aviator teardrops?|shoulder bags?|slips?|sweatpants?|shearling|casual|double breasted|slouch|dress(?:es)?|mary jane|trucker|faux fur|capes?|denim jeans?|logos?|chino khakis?|joggers?|satchels?|maxi|midi|mini|dad sneaker platform|dusters?|anoraks?|raincoats?|full|culotte|platforms?|rectangular|lace tops?|rain|button up|slip dress(?:es)?|cutoff|moccasin|pull ons?|wallet card cases?|blazer dress(?:es)?|strappy|tote bags?|active tracks?|socks?|clutch(?:es)?|pouch(?:es)?|shirt dress(?:es)?|aline|oversized|casual shorts?|single breasted|espadrilles?|cross body bags?|aviator squares?|shift|sweater dress(?:es)?|combats?'
attributes_regex['closure'] = 'back ties?|zip fly with hooks?|belted|ties?|back zips?|button fronts?|wraps?|laces?|open|back zippers?|zippers?|magnetics?|front zip with hooks?|halfway zippers?|hook and loops?|twisteds?|tie fronts?|drawstrings?|open fronts?|belts?|zip fly with buttons?|hooks?|locks?|front zip with buttons?|buttoned fronts?|snaps?|flaps?|buttoned backs?|straps?|front ties?|side zips?|ankle traps?|Hook eyes?|tiebacks?|slingbacks?|14 zips?|front zips?|velcros?|halfway buttoned|zip ups?|tie necks?|clasps?|buckles?'
attributes_regex['color'] = 'beige|black|blue|light brown|dark brown|brown|burgundy|gold|gray|grey|green|multi|navy|neutral|orange|pink|purple|red|silver|teal|white|yellow'
attributes_regex['dry_clean_only'] = 'yes|no'
attributes_regex['embellishment'] = 'patch(?:es)?|trims?|beaded|rhinestones?|fringes?|studs?|crystals?|epaulets?|mesh(?:es)?|sequins?|embroider(?:y|ies)|laces?|ruffles?|buckles?'
attributes_regex['exposure'] = 'closed toes?|open toes?|peep toes?'
attributes_regex['fit'] = 'fitted tailored|semifitted|straight regular|oversized|relaxed|tight'
attributes_regex['height'] = 'thigh high|knee high|over the knees?|flats?|low|high|mid calf|medium|ankles?'
attributes_regex['leg_style'] = 'boyfriends?|flares?|wide legs?|skinny|straight|baggy|boot cuts?|slim|'
attributes_regex['length'] = 'capris?|cropped|short sleeves?|midi mid calf|knee length|high low|sleeveless|cropped above waistlines?|short above the knees?|mid 510|short 35|ankles?|very short 3|maxi at and below calf|regular|shorts?|mid calf|sleeves?|mini mid thigh and above|long below hips?|long sleeves?|short at waistlines?|cap sleeves?|long 10|maxi past calf and below|mid length at hips?|tall'
attributes_regex['neckline'] = 'scoop necks?|halters?|cowl necks?|asymmetricals?|mandarin collars?|one shoulders?|off shoulders?|straplesss?|cold shoulders?|straights?|vnecks?|high necks?|sweethearts?|split necks?|henleys?|keyholes?|drapeds?|mock necks?|collareds?|plunge necks?|turtlenecks?|surplices?|funnels?|shawl collars?|pussy bows?|boat necks?|square necks?|tie necks?|band collars?|jewels?|crew necks?'
attributes_regex['occasion'] = 'works?|night outs?|cold weathers?|weekends?|day tonights?|vacations?|workouts?'
attributes_regex['pattern'] = 'stripe verticals?|plaids?|animals?|monograms?|dots?|logos?|multi prints?|camouflages?|paisleys?|houndstooths?|geometrics?|abstracts?|stripes?|florals?|tiedyes?|graphics?|tropicals?|color blocks?|pinstripes?|stripe horizontals?'
attributes_regex['shape'] = 'wedges?|stilettos?|platform flatforms?|wides?|inward curves?|monograms?|blocks?|raised soles?|ovals?|cones?|rectangles?|kittens?'
attributes_regex['sheer'] = 'sheers?|semisheers?'
attributes_regex['sizing'] = 'regular|tall|plus|maternity'
attributes_regex['strap'] = 'metals?|adjustables?|canvass?|spaghettis?|suedes?|leathers?|synthetics?|pure cottons?|elastic stretchs?|cotton blends?|faux leathers?'
attributes_regex['style'] = 'boho|classic|business casual|casual|glam|edgy|retro|athleisure|modern|romantic|androgynous|'
attributes_regex['toe_style'] = 'pointed toes?|peep toes?|open toes?|closed toes?|round toes?|square toes?'
attributes_regex['trend'] = 'patch|trims?|distressed|coated|clean|raw hems?'
attributes_regex['wash'] = 'acids?|blacks?|raws?|dark blues?|light blues?|denims?|whites?|Greys?|medium blues?'
attributes_regex['width'] = 'skinny up to 1|narrow|medium'


def remove_duplicates(lst):
    '''removes the duplicates in a list'''
    return list(set(lst))

# remove duplicated values in each tags
for key in attributes_regex:
    product[key] = product.lemm_total.str.findall(rf"\b({attributes_regex[key]})\b").apply(remove_duplicates)

# Extract manufacturing location 
product['location'] = product['total'].str.extract(r'made in (?P<location>\w+)')
# Extract materials and its percentage (if any)
product[['material_percent', 'material']] = product['description'].str.extract(r'\b(?P<percent>\d+\%)? ?(?P<materials>linen blends?|metals?|lyocells?|chiffons?|faux furs?|satin charmeuses?|calf hairs?|cashmere blends?|spandexs?|jersey knits?|alpacas?|pure wools?|twills?|synthetics?|pure cottons?|silk blends?|faux leathers?|rayons?|tweeds?|nylons?|canvass?|pony hairs?|viscoses?|wool blends?|denims?|golds?|suedes?|silks?|cupros?|wools?|fleeces?|chenilles?|brasss?|acetates?|pure cashmeres?|acrylics?|crepe de chines?|pure linens?|pure silks?|elastanes?|techs?|raffias?|faux shearlings?|corduroys?|snakeskins?|cotton blends?|shearlings?|linens?|modals?|velvets?|plastics?|mohairs?|beadeds?|chambrays?|polyesters?|sateens?|ramies?|silvers?|leathers?|straws?|patent leathers?|rhinestone|polyamides?|poly chiffon|lurex)\b')

# Extract brand specific words (we looked at the top 30 brands' description and found that these words might be 
# unique to only one/some of them)
brand_features = 'vintage|lightweight|maison louis marie|boy smells|edgy|playful|lush|chic|sustainable|environmentally friendly|need help chat with us|dry clean|hand wash|machine wash|grown in ghana|hand batiking studio|handwoven|french|neoprene|fiberlite|our journey|organic|social and environmental accountability|planet|special collaboration|artist|handmade|classic|vegan|luxe|bestselling|sara ekua todd|opening ceremony|ancient greek|tie dye'
product['brand_specific'] = product['total'].str.findall(rf'\b({brand_features})\b').apply(remove_duplicates)

Tags are saved in list format for "other_tags". We want to convert the list format to words.

other_features = product.columns[product.columns.get_loc('class'):]
other_features

# only keep features columns
features = product.set_index('product_id')
features = features.iloc[:,product.columns.get_loc('lemm_total')-1:]

#convert list format to words
def list_to_word(x):
    if (type(x) != list and type(x)!= str):
        return np.nan
    elif type(x) == list:
        if x == []:
            return np.nan
        else:
            return ' '.join(x)
    elif type(x) == str:
        return x

for column in other_features:
    features[column] = features[column].apply(list_to_word)

features.head(2)

## Create a column for all the features 

- Combined all the tags first and then add the `lemm_total` column in the end, in case the product does not have enough tags.

all_features_except_total = product.columns[product.columns.get_loc('lemm_total')+1:]
all_features_except_total

# combine all the tags
features['all_features'] = np.nan
for column in all_features_except_total:
    features['all_features'] = features['all_features'].map(str) + ' ' + features[column].map(str)

# add lemm_total to combined feature
features['all_features'] = features['all_features'] + ' ' + features['lemm_total']
features['all_features'] = features['all_features'].str.replace('nan', ' ').str.replace('  ', ' ')
features.head(3)

# Final data 

Our final feature data includes the following: 
- **`lemm_total`:** cleaned data by combining `brand_category` + `name` + `details` + `description` + `brand_name`
- **`detailed_category`, `general_category`,`gender`, ......, `brand_specific`** : different types of tags
- **`all_features`:** all tags + `lemm_total`

features.info()

features.to_csv('all_features.csv')
